requirements.txt
pandas
chromadb
google-genai
sentence-transformers
python-dotenv
fastapi
uvicorn[standard]
